+++ src/native/org_catacombae_jfuse_Errno.cpp
@@ -136,6 +136,6 @@ JNIEXPORT jint JNICALL Java_org_catacombae_jfuse_Errno_getNativeErrnoValue
     }
 
     CSLogTraceLeave("jint Java_org_catacombae_jfuse_Errno_getNativeErrnoValue"
-            "(%p, %p, %p): %" PRId32 "", env, cls, errnoName, result);
+            "(%p, %p, %p): %" PRId32 "", env, cls, errnoName, (int32_t)result);
     return result;
 }
